Fixed ordering of categories in the categories page
This CL aims at fixing the ordering of categories in the categories
page. Currently, the order is decided based on what order the categories
are added to the list. We can improve the logic to rely on something
more concrete like a priority value.

Bug: 406517913
Flag: EXEMPT bugfix
Test: Tested by building and installing on local and checking if the
order is correct or not

Change-Id: Ie4ef815753813798a39dc8063704be26fc26b7ea