Seperate view holder for photos sections
This CL aims at having a seperate view holder for photos sections. This
is needed in order to keep the code clean and also to have all photos
section specific behavior in one place. This was also causing a number
of bugs which can now be easily fixed because photos section is one
place.

Bug: 409841415
Flag: com.android.systemui.shared.new_customization_picker_ui
Test: Tested by building and installing picker on local and checking if
i am able to see the carousel or not.

Change-Id: I720b17bee40b752412369944a374dbadcd44c4e5