Attach snap helper only once
This CL aims to make sure that we attach snap helper only once instead
of attaching it each time to the carousel. This is because attaching it
each time causes a crash since the view for carousel gets recycled.

Bug: 410771334, 407021251
Flag: com.android.systemui.shared.new_customization_picker_ui
Test: Tested by tapping on clock style in lock screen and checking if
picker crashes or not. It doesn't crash after the change.

Change-Id: Ib5387ceda7e805b75e1d5c802d6c3dbdf7ab294b