Make sure initial state when secondary screen (1/2)
When the initial state is a secondary screen and the bind is called, we
should make sure the UI is set to the secondary screen before binding
it. This avoids the situation of motion layout transition animation
conflicting with the recycler view items updapte animation.

Test: Manually tested. See bug.
Bug: 412546686
Flag: com.android.systemui.shared.new_customization_picker_ui
Change-Id: Iffb36711029bbbca9a4f47eaa8249954780ffc09