Fix color loading non-stop
Usage of Configuration#diff is incorrect. The new configuration should
be passed into the diff function as parameter rather than the current
configuration because it assumes that the new configuration represents
the delta, and undefined values are ignored. After a restart, the
assetsSeq value is reset in the current configuration and is undefined,
so passing it in causes no asset path diff to be detected.

Flag: com.android.systemui.shared.new_customization_picker_ui
Bug: 422543951
Test: manually verified by restarting and applying color in picker
Change-Id: I35c6f9736ce240ea60ac967f5484d7e8768c130c