Fix hidden small preview on back from full preview
Do not cache arguments in onCreate, since onCreate is not called again
when navigating to a different fragment and then returning to the
current fragment, and the cached value is used instead. To properly
remove a value from arguments after the first time it is read, remove it
after it's consumed and re-read the value from arguemnts each time.

Flag: com.android.systemui.shared.new_customization_picker_ui
Bug: 432382450
Test: manually verified with live, static, and creative wallpapers
Test: manually verified with rotation and unfold on foldables
Test: manually verified by navigating from small to full preview
Change-Id: Ic5541e9782dbd1b79d166acff010f1636ab6e26f