Fix static wallpaper not show
We need to set the crop of the wallpaper after we have the full res
image model. So that we can extract the color of the wallpaper and thus
the workspace surface control can be created.

We need both the wallpaper and workspace surface control to render the
preivew.

Test: Manually tested. See bug
Bug: 423956081
Flag: com.android.wallpaper.refactor_wallpaper_preview_screen_flag
Change-Id: I23e547f9c97bb9b752ad44467b24a23c7c0a8262