Update per interface change
The wallpaper dimming related API has updated to support multi-display
and temporary dimming. Update the calls to match the updated API.

No behavior change is introduced in this change. When the display id is
set to a negative value and the temporary flag is set to false, the API
behaves identical to what it did before.

Bug: 406869685
Test: build
Test: WallpaperManagerServiceTests
Flag: com.android.window.flags.dimming_wallpaper_for_maximized_and_tiled

Change-Id: Id7cb851ae3cec9008e9442f92dc6b2f484842f60