Fix deadlock in Wallpaper Carousel by emitting initial value
Replaced emptyFlow() with flowOf(null) in the default curated photos
interactor. This fixes a bug where the UI would stay in a loading state
forever because the combine operator in the ViewModel was deadlocked
waiting for an emission from the curated photos flow. Providing an
Providing an initial null emission allows the UI to skip this section
skip this section and show default wallpapers.

Bug: b/468384906
Flag: EXEMPT BUGFIX
Test: Tested before and after by installing the AOSP build, and checking
if loading stops and categories page also loads.

Change-Id: Ib6618cfb3794464fc1f18692ca3b7d4d91691c27